edit torch_dtype to dtype

+++ src/lm_saes/backend/language_model.py
@@ -291,7 +291,7 @@ class HuggingFaceLanguageModel(LanguageModel):
             cfg.model_name if cfg.model_from_pretrained_path is None else cfg.model_from_pretrained_path,
             cache_dir=cfg.cache_dir,
             local_files_only=cfg.local_files_only,
-            torch_dtype=cfg.dtype,
+            dtype=cfg.dtype,
             trust_remote_code=True,
         ).to(self.device)  # type: ignore
